Re-order auth functions to their relevant file
